+++ packages/wiz-ui-react/src/components/base/inputs/date-range-picker/components/date-range-picker.tsx
@@ -107,6 +107,19 @@ const DateRangePicker: FC<Props> = ({
     );
   };
 
+  const initiaizeRightCalendarDate = () => {
+    const [start, end] = [dateRange.start, dateRange.end];
+    if (end) {
+      setRightCalendarDate(new Date(end));
+    } else if (start) {
+      setRightCalendarDate(
+        new Date(start.getFullYear(), start.getMonth() + 1, 1)
+      );
+    } else {
+      setRightCalendarDate(new Date());
+    }
+  };
+
   const selectedDates = useMemo(() => {
     const getDateStatus = (
       date: Date,
@@ -189,6 +202,7 @@ const DateRangePicker: FC<Props> = ({
   })();
 
   const onClose = () => {
+    initiaizeRightCalendarDate();
     setTempDateRange(dateRange);
     setIsOpen(false);
   };
@@ -225,25 +239,30 @@ const DateRangePicker: FC<Props> = ({
           <div className={styles.popupCalendarCancelIconStyle}>
             <WizIcon size="xl2" color="gray.500" icon={WizICalendar} />
           </div>
-          <div
+
+          <span
             className={
               styles.inputTextStyle[
-                dateRange.start && !disabled ? "selected" : "default"
+                (disabled && "disabled") ||
+                  (dateRange.start && "selected") ||
+                  "default"
               ]
             }
           >
             {dateRange.start ? formatDate(dateRange.start) : "開始日"}
-          </div>
-          <div className={styles.separatorStyle}>-</div>
-          <div
+          </span>
+          <span className={styles.separatorStyle}>-</span>
+          <span
             className={
               styles.inputTextStyle[
-                dateRange.end && !disabled ? "selected" : "default"
+                (disabled && "disabled") ||
+                  (dateRange.end && "selected") ||
+                  "default"
               ]
             }
           >
             {dateRange.end ? formatDate(dateRange.end) : "終了日"}
-          </div>
+          </span>
         </WizHStack>
 
         {cancelButtonVisible && (
@@ -260,7 +279,7 @@ const DateRangePicker: FC<Props> = ({
       </button>
       <WizPopup
         isOpen={!disabled && isOpen}
-        onClose={() => setIsOpen(false)}
+        onClose={onClose}
         isDirectionFixed={isDirectionFixed}
         anchorElement={anchor}
       >