+++ packages/wiz-ui-next/CHANGELOG.md
@@ -1,5 +1,93 @@
 # @wizleap-inc/wiz-ui-next
 
+## 2.45.1
+
+### Patch Changes
+
+- [#1397](https://github.com/Wizleap-Inc/wiz-ui/pull/1397) [`f34a91b28cf53af8fa0c712c84dd33b053c5a44c`](https://github.com/Wizleap-Inc/wiz-ui/commit/f34a91b28cf53af8fa0c712c84dd33b053c5a44c) Thanks [@k-tada](https://github.com/k-tada)! - FormControl の nowrap 実装を修正
+
+## 2.45.0
+
+### Minor Changes
+
+- [#1336](https://github.com/Wizleap-Inc/wiz-ui/pull/1336) [`5ed4d076b472d9d441b752dc007d9ca0e14ec4bb`](https://github.com/Wizleap-Inc/wiz-ui/commit/5ed4d076b472d9d441b752dc007d9ca0e14ec4bb) Thanks [@RyushiAok](https://github.com/RyushiAok)! - chore: update disabled styles
+
+### Patch Changes
+
+- Updated dependencies [[`5ed4d076b472d9d441b752dc007d9ca0e14ec4bb`](https://github.com/Wizleap-Inc/wiz-ui/commit/5ed4d076b472d9d441b752dc007d9ca0e14ec4bb)]:
+  - @wizleap-inc/wiz-ui-styles@0.35.0
+
+## 2.44.2
+
+### Patch Changes
+
+- [#1392](https://github.com/Wizleap-Inc/wiz-ui/pull/1392) [`1a1753205d42d02853b46491cfc299fd7ce5cb10`](https://github.com/Wizleap-Inc/wiz-ui/commit/1a1753205d42d02853b46491cfc299fd7ce5cb10) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Fix: nowrap form control
+
+- [#1391](https://github.com/Wizleap-Inc/wiz-ui/pull/1391) [`05d4af989a31a93a8f01919346ebd5980ce1d87f`](https://github.com/Wizleap-Inc/wiz-ui/commit/05d4af989a31a93a8f01919346ebd5980ce1d87f) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Fix: tooltip を dropdown より上に表示する
+
+- Updated dependencies [[`05d4af989a31a93a8f01919346ebd5980ce1d87f`](https://github.com/Wizleap-Inc/wiz-ui/commit/05d4af989a31a93a8f01919346ebd5980ce1d87f)]:
+  - @wizleap-inc/wiz-ui-constants@0.33.0
+  - @wizleap-inc/wiz-ui-styles@0.34.2
+
+## 2.44.1
+
+### Patch Changes
+
+- [#1383](https://github.com/Wizleap-Inc/wiz-ui/pull/1383) [`9286a5810c7e1d0a88065254b246f4efa05efd31`](https://github.com/Wizleap-Inc/wiz-ui/commit/9286a5810c7e1d0a88065254b246f4efa05efd31) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Fix(number-input): height
+
+- Updated dependencies [[`9286a5810c7e1d0a88065254b246f4efa05efd31`](https://github.com/Wizleap-Inc/wiz-ui/commit/9286a5810c7e1d0a88065254b246f4efa05efd31)]:
+  - @wizleap-inc/wiz-ui-styles@0.34.1
+
+## 2.44.0
+
+### Minor Changes
+
+- [#1379](https://github.com/Wizleap-Inc/wiz-ui/pull/1379) [`d0e8556a47433c4e40226fe171b638e593103061`](https://github.com/Wizleap-Inc/wiz-ui/commit/d0e8556a47433c4e40226fe171b638e593103061) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat: WizNumberInput
+
+### Patch Changes
+
+- Updated dependencies [[`d0e8556a47433c4e40226fe171b638e593103061`](https://github.com/Wizleap-Inc/wiz-ui/commit/d0e8556a47433c4e40226fe171b638e593103061)]:
+  - @wizleap-inc/wiz-ui-constants@0.32.0
+  - @wizleap-inc/wiz-ui-styles@0.34.0
+
+## 2.43.0
+
+### Minor Changes
+
+- [#1377](https://github.com/Wizleap-Inc/wiz-ui/pull/1377) [`4f2aad5fbbf8ac28f0d63c3a6fded155761fe5a8`](https://github.com/Wizleap-Inc/wiz-ui/commit/4f2aad5fbbf8ac28f0d63c3a6fded155761fe5a8) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(icon): filled counter
+
+### Patch Changes
+
+- Updated dependencies [[`4f2aad5fbbf8ac28f0d63c3a6fded155761fe5a8`](https://github.com/Wizleap-Inc/wiz-ui/commit/4f2aad5fbbf8ac28f0d63c3a6fded155761fe5a8)]:
+  - @wizleap-inc/wiz-ui-constants@0.31.0
+  - @wizleap-inc/wiz-ui-styles@0.33.1
+
+## 2.42.0
+
+### Minor Changes
+
+- [#1375](https://github.com/Wizleap-Inc/wiz-ui/pull/1375) [`ff7f53984d4a7f6a9315bb9d95f1fc9f5a268a7e`](https://github.com/Wizleap-Inc/wiz-ui/commit/ff7f53984d4a7f6a9315bb9d95f1fc9f5a268a7e) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Chore(infomation-panel): update close button position
+
+- [#1374](https://github.com/Wizleap-Inc/wiz-ui/pull/1374) [`576c3c1b9795136d3df465a989bb4d8eef520b5b`](https://github.com/Wizleap-Inc/wiz-ui/commit/576c3c1b9795136d3df465a989bb4d8eef520b5b) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(icons): add counter, tunes, cloud_upload
+
+### Patch Changes
+
+- Updated dependencies [[`ff7f53984d4a7f6a9315bb9d95f1fc9f5a268a7e`](https://github.com/Wizleap-Inc/wiz-ui/commit/ff7f53984d4a7f6a9315bb9d95f1fc9f5a268a7e), [`576c3c1b9795136d3df465a989bb4d8eef520b5b`](https://github.com/Wizleap-Inc/wiz-ui/commit/576c3c1b9795136d3df465a989bb4d8eef520b5b)]:
+  - @wizleap-inc/wiz-ui-styles@0.33.0
+  - @wizleap-inc/wiz-ui-constants@0.30.0
+
+## 2.41.1
+
+### Patch Changes
+
+- [#1372](https://github.com/Wizleap-Inc/wiz-ui/pull/1372) [`c041edba88dbd007b3af58a541c30ba090d87898`](https://github.com/Wizleap-Inc/wiz-ui/commit/c041edba88dbd007b3af58a541c30ba090d87898) Thanks [@ichi-h](https://github.com/ichi-h)! - fix(wiz-ui-next): ジェネリクスの型がうまく生成できない問題の解消
+
+## 2.41.0
+
+### Minor Changes
+
+- [#1345](https://github.com/Wizleap-Inc/wiz-ui/pull/1345) [`73c5b87cfc3c75239de8814e97a26c42d1b3fcff`](https://github.com/Wizleap-Inc/wiz-ui/commit/73c5b87cfc3c75239de8814e97a26c42d1b3fcff) Thanks [@RyushiAok](https://github.com/RyushiAok)! - Feat(selects): ジェネリックコンポーネント化
+
 ## 2.40.0
 
 ### Minor Changes